Augmented the file to include writing and reading routines. The routines write to and read from a contiguous chunk in the dataset, which can be in an arbitrary location. Reading does not care of what that chunk of data looks like. In writing, care needs to be taken to not have overlaps of the data chunks on the disk.
Writing/reading operations are collective, and all processes in the
communicator need to participate. Should a process need to write/read no
data, the count can be set to zero (and the offset can be zero).

The routines are designed such that writing and reading can happen in a way
that all processes write/read a portion of the entire dataset. This does
not have to be the case. For either writing or reading, these routines can
be wrapped in sweeps, and be used to "throttle" the write/read operations
for reasons that may concern efficiency of the underlying parallel file
system or memory management on behalf of the application.
IN 2019/04/29